Solved assertion error. A refresh() method must be called before retrieveing account data from database
The reason why the account balance assertions failed was because the data was not refreshed when the account.getBalance() method was called.
The ActiveJDBC BASE class to which Account class is extended to has a refresh() method.
Calling this method before retrieving the balance with getString method from the database solved the problem.

This commit has all the comments and consol output lines included, but will be removed by next commit.